+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 3.0.0
+- MockFactory changed:
+    -  Passing constructor parameters for instantiate a class with constructor parameters.
+    - Define global variable mock for accessing mocks.